+++ node/middlewares/checkout.ts
@@ -22,7 +22,7 @@ export async function digitalRiverCreateCheckout(
   next: () => Promise<unknown>
 ) {
   const {
-    clients: { apps, digitalRiver, logistics, orderForm },
+    clients: { apps, digitalRiver, orderForm },
     req,
     vtex: { logger },
   } = ctx
@@ -74,33 +74,34 @@ export async function digitalRiverCreateCheckout(
   }
 
   const items = []
-  const docks = []
+  // const docks = []
 
-  for (const logisticsInfo of orderFormData.shippingData.logisticsInfo) {
-    const { selectedSla, slas } = logisticsInfo
-    const [{ dockId }] = slas.find(
-      ({ name }: { name: string }) => name === selectedSla
-    ).deliveryIds
+  // for (const logisticsInfo of orderFormData.shippingData.logisticsInfo) {
+  //   const { selectedSla, slas } = logisticsInfo
+  //   const [{ dockId }] = slas.find(
+  //     ({ name }: { name: string }) => name === selectedSla
+  //   ).deliveryIds
 
-    // eslint-disable-next-line no-await-in-loop
-    let dockInfo = await logistics.getDocksById(dockId)
+  //   // eslint-disable-next-line no-await-in-loop
+  //   let dockInfo = await logistics.getDocksById(dockId)
 
-    if (
-      !dockInfo.address?.city ||
-      !dockInfo.address?.postalCode ||
-      !dockInfo.address?.country?.acronym
-    ) {
-      logger.warn({
-        message: 'DigitalRiverCreateCheckout-dockAddressMisconfiguration',
-        dockInfo,
-      })
-      dockInfo = ''
-    }
+  //   if (
+  //     !dockInfo?.address?.city ||
+  //     !dockInfo?.address?.postalCode ||
+  //     !dockInfo?.address?.country?.acronym
+  //   ) {
+  //     logger.warn({
+  //       message: 'DigitalRiverCreateCheckout-dockAddressMisconfiguration',
+  //       dockInfo,
+  //     })
+  //     dockInfo = ''
+  //   }
 
-    docks.push(dockInfo)
-  }
+  //   docks.push(dockInfo)
+  // }
 
-  for (const [index, item] of orderFormData.items.entries()) {
+  // for (const [index, item] of orderFormData.items.entries()) {
+  for (const item of orderFormData.items) {
     let discountPrice = 0
 
     for (const priceTag of item.priceTags) {
@@ -115,7 +116,7 @@ export async function digitalRiverCreateCheckout(
       }
     }
 
-    const dock = docks[index]
+    // const dock = docks[index]
 
     const newItem: CheckoutItem = {
       skuId: item.id,
@@ -127,18 +128,18 @@ export async function digitalRiverCreateCheckout(
             quantity: item.quantity,
           }
         : undefined,
-      ...(!!dock && {
-        shipFrom: {
-          address: {
-            line1: dock.address.street || 'Unknown',
-            line2: dock.address.complement || '',
-            city: dock.address.city,
-            postalCode: dock.address.postalCode,
-            state: dock.address.state || '',
-            country: convertIso3To2(dock.address.country.acronym),
-          },
-        },
-      }),
+      // ...(!!dock && {
+      //   shipFrom: {
+      //     address: {
+      //       line1: dock.address.street || 'Unknown',
+      //       line2: dock.address.complement || '',
+      //       city: dock.address.city,
+      //       postalCode: dock.address.postalCode,
+      //       state: dock.address.state || '',
+      //       country: convertIso3To2(dock.address.country.acronym),
+      //     },
+      //   },
+      // }),
     }
 
     items.push(newItem)
@@ -153,6 +154,16 @@ export async function digitalRiverCreateCheckout(
     currency: orderFormData?.storePreferencesData?.currencyCode ?? 'USD',
     taxInclusive: true,
     email: orderFormData.clientProfileData?.email ?? '',
+    shipFrom: {
+      address: {
+        line1: orderFormData.shippingData?.address?.street || 'Unknown',
+        line2: orderFormData.shippingData?.address?.complement || '',
+        city: orderFormData.shippingData?.address?.city || 'Unknown',
+        state: orderFormData.shippingData?.address?.state || '',
+        postalCode: orderFormData.shippingData?.address?.postalCode || '',
+        country: shippingCountry,
+      },
+    },
     shipTo: {
       name:
         orderFormData.clientProfileData?.firstName &&